fix(runtime): 修复 packet trace lint 告警

+++ include/yunlink/core/runtime_config.hpp
@@ -76,7 +76,7 @@ struct RuntimeConfig {
     uint32_t security_key_epoch = 1;
     bool packet_trace_enabled = false;
     size_t packet_trace_max_records = 500;
-    size_t packet_trace_max_total_bytes = 8 * 1024 * 1024;
+    size_t packet_trace_max_total_bytes = static_cast<size_t>(8) * 1024 * 1024;
     size_t packet_trace_raw_preview_bytes = 4096;
     size_t packet_trace_payload_preview_bytes = 4096;
 };
